+++ tests/integration/12_config_protect.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
-export LUET_NOLOCK=true
-export LUET_BUILD=luet-build
-export LUET=luet
+testsourcedir=$(dirname "${BASH_SOURCE[0]}")
+source ${testsourcedir}/_common.sh
 
 oneTimeSetUp() {
-export tmpdir="$(mktemp -d)"
+  export tmpdir="${TEST_TMPDIR:-$(mktemp -d)}"
 }
 
 oneTimeTearDown() {
+  if [ -z "${SKIP_CLEAN}" ] ; then
     rm -rf "$tmpdir"
+  fi
 }
 
 testBuild() {
@@ -37,7 +38,6 @@ testRepo() {
 }
 
 testConfig() {
-
     mkdir $tmpdir/testrootfs/etc/luet/config.protect.d -p
 
     cat <<EOF > $tmpdir/testrootfs/etc/luet/config.protect.d/conf1.yml
@@ -51,10 +51,8 @@ general:
   debug: true
 system:
   rootfs: $tmpdir/testrootfs
-  database_path: "/"
+  database_path: "/var/cache/luet"
   database_engine: "boltdb"
-config_protect_confdir:
-    - /etc/luet/config.protect.d
 config_from_host: false
 repositories:
    - name: "main"
@@ -88,7 +86,7 @@ testInstall() {
 
 
 testUnInstall() {
-    $LUET uninstall -y --full --config $tmpdir/luet.yaml test/a
+    $LUET uninstall -y --config $tmpdir/luet.yaml test/a
     installst=$?
     assertEquals 'uninstall test successfully' "$installst" "0"
     assertTrue 'package uninstalled' "[ ! -e '$tmpdir/testrootfs/c' ]"
@@ -99,7 +97,7 @@ testUnInstall() {
 
 
 testCleanup() {
-    $LUET cleanup --config $tmpdir/luet.yaml
+    $LUET cleanup --config $tmpdir/luet.yaml --purge-repos
     installst=$?
     assertEquals 'install test successfully' "$installst" "0"
     assertTrue 'package installed' "[ ! -e '$tmpdir/testrootfs/packages/a-test-1.0.package.tar.gz' ]"
